[ty] Move shared call-unwrapping fixes into ruff_python_edits
