+++ .github/workflows/ci.yaml
@@ -241,6 +241,7 @@ jobs:
             ':crates/ruff_db/**' \
             ':crates/ruff_annotate_snippets/**' \
             ':crates/ruff_python_ast/**' \
+            ':crates/ruff_python_edits/**' \
             ':crates/ruff_python_parser/**' \
             ':crates/ruff_python_trivia/**' \
             ':crates/ruff_source_file/**' \
