+++ .github/workflows/typing_conformance.yaml
@@ -13,6 +13,7 @@ on:
       - "!crates/ty_wasm/**"
       - "crates/ruff_db"
       - "crates/ruff_python_ast"
+      - "crates/ruff_python_edits/**"
       - "crates/ruff_python_parser"
       - "scripts/conformance.py"
       - ".github/workflows/typing_conformance.yaml"
