+++ Cargo.toml
@@ -30,6 +30,7 @@ ruff_notebook = { version = "0.0.11", path = "crates/ruff_notebook" }
 ruff_options_metadata = { version = "0.0.11", path = "crates/ruff_options_metadata" }
 ruff_python_ast = { version = "0.0.11", path = "crates/ruff_python_ast" }
 ruff_python_codegen = { version = "0.0.11", path = "crates/ruff_python_codegen" }
+ruff_python_edits = { version = "0.0.11", path = "crates/ruff_python_edits" }
 ruff_python_formatter = { version = "0.0.11", path = "crates/ruff_python_formatter" }
 ruff_python_importer = { version = "0.0.11", path = "crates/ruff_python_importer" }
 ruff_python_index = { version = "0.0.11", path = "crates/ruff_python_index" }
