+++ crates/ruff_linter/src/rules/ruff/rules/unnecessary_cast_to_int.rs
@@ -1,6 +1,6 @@
 use ruff_macros::{ViolationMetadata, derive_message_formats};
-use ruff_python_ast::helpers::unwrapped_call_argument;
 use ruff_python_ast::{Arguments, Expr, ExprCall};
+use ruff_python_edits::unwrapped_call_argument;
 use ruff_python_semantic::SemanticModel;
 use ruff_python_semantic::analyze::type_inference::{NumberLike, PythonType, ResolvedPythonType};
 use ruff_text_size::{Ranged, TextRange};
