+++ crates/ruff_python_edits/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "ruff_python_edits"
+version = "0.0.11"
+description = "This is an internal component crate of Ruff"
+authors = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+homepage = { workspace = true }
+documentation = { workspace = true }
+repository = { workspace = true }
+license = { workspace = true }
+
+[dependencies]
+ruff_python_ast = { workspace = true }
+ruff_source_file = { workspace = true }
+ruff_text_size = { workspace = true }
+
+[dev-dependencies]
+ruff_python_parser = { workspace = true }
+
+[lints]
+workspace = true
+
+[lib]
+doctest = false
