+++ crates/ruff_python_edits/README.md
@@ -1,0 +1,12 @@
+<!-- This file is generated. DO NOT EDIT -->
+
+# ruff_python_edits
+
+This crate is an internal component of [Ruff](https://crates.io/crates/ruff). The Rust API exposed
+here is unstable and will have frequent breaking changes.
+
+This version (0.0.11) is a component of [Ruff 0.16.5](https://crates.io/crates/ruff/0.16.5). The
+source can be found [here](https://github.com/astral-sh/ruff/blob/0.16.5/crates/ruff_python_edits).
+
+See Ruff's [crate versioning policy](https://docs.astral.sh/ruff/versioning/#crate-versioning) for
+details on versioning.
