+++ crates/ty_python_semantic/src/types/function.rs
@@ -60,8 +60,8 @@ use ruff_db::parsed::{ParsedModuleRef, parsed_module};
 use ruff_db::source::source_text;
 use ruff_diagnostics::{Edit, Fix};
 use ruff_python_ast::find_node::covering_node;
-use ruff_python_ast::helpers::unwrapped_call_argument;
 use ruff_python_ast::{self as ast, ParameterWithDefault};
+use ruff_python_edits::unwrapped_call_argument;
 use ruff_text_size::Ranged;
 use salsa::plumbing::AsId;
 use ty_module_resolver::{ImportingFile, KnownModule, ModuleName, file_to_module, resolve_module};
